Updates for MaterialData changes Try to generator chunks

+++ src/main/java/org/getspout/server/util/ChannelBufferUtils.java
@@ -106,7 +106,7 @@ public final class ChannelBufferUtils {
 					int id = buf.readShort();
 					int count = buf.readByte();
 					short damage = buf.readShort();
-					ItemStack item = new ItemStack(MaterialData.getMaterial(id), count, damage);
+					ItemStack item = new ItemStack(MaterialData.getMaterial((short)id), count, damage);
 					parameters.add(new Parameter<ItemStack>(type, index, item));
 					break;
 			}
